Refactoring how nodes and ways are added
When adding nodes to a WeightedNodeSet, alternative coordinates
can be assigned to this added node. This is relevant if the added
node represents a group of nodes, and the coordinates shall
represent the center of this group.
Dumping information stored in the WeightedNodeSet takes those
shifted coordinates into account.

When adding ways, it is checked if this way is closed. In this
case, the way is collapsed into a central point and the way's
first node with different coordinates is used as a representative.

When adding relations, "inner" roles are skipped as they are usually
not very helpful.