Exploiting unbalanced structure of tree for performance improvement
It is an observation that out of the 64-bit space for ids, most used ids
are small, i.e. have their most significant bits set to zero. This
results in that the id tree will be very unbalanced, as for the first
levels the 0th child will be used.
To avoid costly traversals of a tree that is unbalanced, here the id is
checked if the bits used for the up to eight first traversal steps are
all set to zero. If true, a pre-recorded path ('zeroPath') will be used
that guides the search function faster through the tree than a 'regular'
tree traversal would do.
Initial tests showed a 25% decrease in CPU time if using this function.