Detection of roads' closeness to cities/towns
Refactoring the detection of roads' closeness to cities or towns.
Roads which are regional (number >= 500) but for which no region
code is known are now detected and handled correctly.
Also, an attempt is made to identify the most likely regional code
for such road; the result (region code) is returned to the caller
to update its data.