Refactoring generation of word combinations into dedicated function
The generation of word combinations, i.e. short sequences of words that may
make up an useful expression, has been refactored from other functions into
a dedicated one, avoiding code duplication.
The function is configurable in what is the maximum number of words to be
combined and if the result may contain duplicates or if each combination
must be unique.