Determining result coordinates split into specialized code
As an ongoing process, specialized algorithms will be used
to determined an input text's location. This commit includes
some basic refactoring for the code in general and as the
first application the detection and cross-referencing of
roads (European, national, regional) and towns/cities;
both which have been identified in the input text.