Getting representative node from way or relation
In several situations, it is necessary to retrieve a
'representative' node for a way or relation. This node
may be used to retrieve a coordinate to learn where
approximately the way is located.

The function introduced in this commit allows to
resolve a given relation or way into a node.
Resolution is done by starting from a given way or
relation, taking its n/2-th member (n in total) as the
relation, way, or node for the next iteration. The
will eventually terminate if the picked member is a node
or the current way or relation has no members to pick
from.