Cross-referencing admin regions and word combinations
Cross-referencing all known administrative regions
(municipalities, counties, ...) and all matching
nodes/ways/relations for each item taken from a list
of word combinations (usually 1..3 words combined).

The list of matches will be sorted by the number of
spaces in the word combination that returned matches,
where word combinations with more spaces are preferred
(it is assumed that more words in a combination make
the combination more specific and as such a better
hit).