Refactoring generation of word combinations
The generation of word combinations now takes added grammar cases,
such as the addition of 'Skaraborg' if 'Skaraborgs' was part of the
input, into consideration. Added grammar cases may only be the
last word of a word combination.

For example: The input 'Skaraborgs tingsrätt' may generate the
combinations 'Skaraborg', 'Skaraborgs', 'tingsrätt', and
'Skaraborgs tingsrätt', but not 'Skaraborg tingsrätt'.