Not storing nodes' names if nodes represent counties or municipalities
At the time of this commit, many counties and municipalities in the
OpenStreetMap database are represented both by nodes (e.g. at the
cities' center) as well as by relations (modern approach) recording
the administrative regions' boundaries.

This commit will prevent recording the names of counties or
municipalities represented by nodes. Thus, searching for e.g.
'Sundsvalls kommun' will not result in both the node and the
relation representing this administrative region. This is
relevant for the 'administrative region' test.