Map analysis class to find crossing roads
Adding a new class MapAnalysis. Its first and so far only purpose is to
find crossing roads.
This is done by going through the given input text and identifying two
road names that are adjecent to each other. All road names that got
identified are then checked if they share a common node which is then
assumed to be the crossing between both roads.